requirements.txt
openai
streamlit>=1.30.0
PyPDF2>=3.0.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
numpy>=1.24.0